use twig loader -> isFresh method to check template cache time.

+++ src/ActionKit/ActionRunner.php
@@ -227,14 +227,16 @@ class ActionRunner
 
         if ( isset( $this->dynamicActions[ $class ] ) ) {
             $args = $this->dynamicActions[ $class ];
-            $gen = new ActionGenerator;
-            $code = $gen->generate($class, $args['template'], $args['variables']);
             $cacheFile = $this->getClassCacheFile($class);
-            // if ( ! file_exists($cacheFile) ) {
+
+            $gen = new ActionGenerator;
+            $loader = $gen->getTwigLoader();
+            if (  ! file_exists($cacheFile) || ! $loader->isFresh($args['template'], filemtime($cacheFile) ) ) {
+                $code = $gen->generate($class, $args['template'], $args['variables']);
                 if ( false === file_put_contents($cacheFile, $code) ) {
                     throw new Exception("Can not write action class cache file: $cacheFile");
                 }
-            // }
+            }
             require $cacheFile;
             return new $class($args);
         }
